refactor: convert to in-memory frequency generation for serverless
- Remove generated JSON files (notes-extended.json, chords-enhanced.json)
- Generate frequencies and chord voicings at runtime on the server
- Remove build scripts that wrote to filesystem
- Optimize for serverless deployment (Vercel)
- Maintain all functionality with zero filesystem writes

The data is now computed in chordsData.server.ts and passed to the frontend via SvelteKit's load pattern.
