--- src/lib/data/chords-enhanced.json
@@ -1,546 +1,0 @@
-{
-	"C": {
-		"standard": [
-			"C4",
-			"E4",
-			"G4"
-		],
-		"spread": [
-			"C3",
-			"E4",
-			"G5"
-		],
-		"rich": [
-			"C2",
-			"C3",
-			"E4",
-			"G4",
-			"C5"
-		],
-		"bass": [
-			"C2",
-			"C4",
-			"E4",
-			"G4"
-		],
-		"inversions": {
-			"first": [
-				"E4",
-				"G4",
-				"C5"
-			],
-			"second": [
-				"G4",
-				"C5",
-				"E5"
-			]
-		}
-	},
-	"Cm": {
-		"standard": [
-			"C4",
-			"Eb4",
-			"G4"
-		],
-		"spread": [
-			"C3",
-			"Eb4",
-			"G5"
-		],
-		"rich": [
-			"C2",
-			"C3",
-			"Eb4",
-			"G4",
-			"C5"
-		],
-		"bass": [
-			"C2",
-			"C4",
-			"Eb4",
-			"G4"
-		]
-	},
-	"Db": {
-		"standard": [
-			"Db4",
-			"F4",
-			"Ab4"
-		],
-		"spread": [
-			"Db3",
-			"F4",
-			"Ab5"
-		],
-		"rich": [
-			"Db2",
-			"Db3",
-			"F4",
-			"Ab4",
-			"Db5"
-		],
-		"bass": [
-			"Db2",
-			"Db4",
-			"F4",
-			"Ab4"
-		]
-	},
-	"Csm": {
-		"standard": [
-			"Db4",
-			"E4",
-			"Ab4"
-		],
-		"spread": [
-			"Db3",
-			"E4",
-			"Ab5"
-		],
-		"rich": [
-			"Db2",
-			"Db3",
-			"E4",
-			"Ab4",
-			"Db5"
-		]
-	},
-	"D": {
-		"standard": [
-			"D4",
-			"Gb4",
-			"A4"
-		],
-		"spread": [
-			"D3",
-			"Gb4",
-			"A5"
-		],
-		"rich": [
-			"D2",
-			"D3",
-			"Gb4",
-			"A4",
-			"D5"
-		],
-		"bass": [
-			"D2",
-			"D4",
-			"Gb4",
-			"A4"
-		]
-	},
-	"Dm": {
-		"standard": [
-			"D4",
-			"F4",
-			"A4"
-		],
-		"spread": [
-			"D3",
-			"F4",
-			"A5"
-		],
-		"rich": [
-			"D2",
-			"D3",
-			"F4",
-			"A4",
-			"D5"
-		]
-	},
-	"Eb": {
-		"standard": [
-			"Eb4",
-			"G4",
-			"Bb4"
-		],
-		"spread": [
-			"Eb3",
-			"G4",
-			"Bb5"
-		],
-		"rich": [
-			"Eb2",
-			"Eb3",
-			"G4",
-			"Bb4",
-			"Eb5"
-		],
-		"bass": [
-			"Eb2",
-			"Eb4",
-			"G4",
-			"Bb4"
-		]
-	},
-	"Dsm": {
-		"standard": [
-			"Eb4",
-			"Gb4",
-			"Bb4"
-		],
-		"spread": [
-			"Eb3",
-			"Gb4",
-			"Bb5"
-		]
-	},
-	"Ebm": {
-		"standard": [
-			"Eb4",
-			"Gb4",
-			"Bb4"
-		],
-		"spread": [
-			"Eb3",
-			"Gb4",
-			"Bb5"
-		],
-		"rich": [
-			"Eb2",
-			"Eb3",
-			"Gb4",
-			"Bb4",
-			"Eb5"
-		]
-	},
-	"E": {
-		"standard": [
-			"E4",
-			"Ab4",
-			"B4"
-		],
-		"spread": [
-			"E3",
-			"Ab4",
-			"B5"
-		],
-		"rich": [
-			"E2",
-			"E3",
-			"Ab4",
-			"B4",
-			"E5"
-		],
-		"bass": [
-			"E2",
-			"E4",
-			"Ab4",
-			"B4"
-		]
-	},
-	"Em": {
-		"standard": [
-			"E4",
-			"G4",
-			"B4"
-		],
-		"spread": [
-			"E3",
-			"G4",
-			"B5"
-		],
-		"rich": [
-			"E2",
-			"E3",
-			"G4",
-			"B4",
-			"E5"
-		]
-	},
-	"F": {
-		"standard": [
-			"F4",
-			"A4",
-			"C5"
-		],
-		"spread": [
-			"F3",
-			"A4",
-			"C6"
-		],
-		"rich": [
-			"F2",
-			"F3",
-			"A4",
-			"C5",
-			"F5"
-		],
-		"bass": [
-			"F2",
-			"F4",
-			"A4",
-			"C5"
-		]
-	},
-	"Fm": {
-		"standard": [
-			"F4",
-			"Ab4",
-			"C5"
-		],
-		"spread": [
-			"F3",
-			"Ab4",
-			"C6"
-		],
-		"rich": [
-			"F2",
-			"F3",
-			"Ab4",
-			"C5",
-			"F5"
-		]
-	},
-	"Gb": {
-		"standard": [
-			"Gb4",
-			"Bb4",
-			"Db5"
-		],
-		"spread": [
-			"Gb3",
-			"Bb4",
-			"Db6"
-		],
-		"rich": [
-			"Gb2",
-			"Gb3",
-			"Bb4",
-			"Db5",
-			"Gb5"
-		],
-		"bass": [
-			"Gb2",
-			"Gb4",
-			"Bb4",
-			"Db5"
-		]
-	},
-	"Fsm": {
-		"standard": [
-			"Gb4",
-			"A4",
-			"Db5"
-		],
-		"spread": [
-			"Gb3",
-			"A4",
-			"Db6"
-		]
-	},
-	"G": {
-		"standard": [
-			"G4",
-			"B4",
-			"D5"
-		],
-		"spread": [
-			"G3",
-			"B4",
-			"D6"
-		],
-		"rich": [
-			"G2",
-			"G3",
-			"B4",
-			"D5",
-			"G5"
-		],
-		"bass": [
-			"G2",
-			"G4",
-			"B4",
-			"D5"
-		]
-	},
-	"Gm": {
-		"standard": [
-			"G4",
-			"Bb4",
-			"D5"
-		],
-		"spread": [
-			"G3",
-			"Bb4",
-			"D6"
-		],
-		"rich": [
-			"G2",
-			"G3",
-			"Bb4",
-			"D5",
-			"G5"
-		]
-	},
-	"Ab": {
-		"standard": [
-			"Ab4",
-			"C5",
-			"Eb5"
-		],
-		"spread": [
-			"Ab3",
-			"C5",
-			"Eb6"
-		],
-		"rich": [
-			"Ab2",
-			"Ab3",
-			"C5",
-			"Eb5",
-			"Ab5"
-		],
-		"bass": [
-			"Ab2",
-			"Ab4",
-			"C5",
-			"Eb5"
-		]
-	},
-	"Gsm": {
-		"standard": [
-			"Ab4",
-			"B4",
-			"Eb5"
-		],
-		"spread": [
-			"Ab3",
-			"B4",
-			"Eb6"
-		]
-	},
-	"A": {
-		"standard": [
-			"A4",
-			"Db5",
-			"E5"
-		],
-		"spread": [
-			"A3",
-			"Db5",
-			"E6"
-		],
-		"rich": [
-			"A2",
-			"A3",
-			"Db5",
-			"E5",
-			"A5"
-		],
-		"bass": [
-			"A2",
-			"A4",
-			"Db5",
-			"E5"
-		]
-	},
-	"Am": {
-		"standard": [
-			"A4",
-			"C5",
-			"E5"
-		],
-		"spread": [
-			"A3",
-			"C5",
-			"E6"
-		],
-		"rich": [
-			"A2",
-			"A3",
-			"C5",
-			"E5",
-			"A5"
-		]
-	},
-	"Bb": {
-		"standard": [
-			"Bb4",
-			"D5",
-			"F5"
-		],
-		"spread": [
-			"Bb3",
-			"D5",
-			"F6"
-		],
-		"rich": [
-			"Bb2",
-			"Bb3",
-			"D5",
-			"F5",
-			"Bb5"
-		],
-		"bass": [
-			"Bb2",
-			"Bb4",
-			"D5",
-			"F5"
-		]
-	},
-	"Bbm": {
-		"standard": [
-			"Bb4",
-			"Db5",
-			"F5"
-		],
-		"spread": [
-			"Bb3",
-			"Db5",
-			"F6"
-		],
-		"rich": [
-			"Bb2",
-			"Bb3",
-			"Db5",
-			"F5",
-			"Bb5"
-		]
-	},
-	"B": {
-		"standard": [
-			"B4",
-			"Eb5",
-			"Gb5"
-		],
-		"spread": [
-			"B3",
-			"Eb5",
-			"Gb6"
-		],
-		"rich": [
-			"B2",
-			"B3",
-			"Eb5",
-			"Gb5",
-			"B5"
-		],
-		"bass": [
-			"B2",
-			"B4",
-			"Eb5",
-			"Gb5"
-		]
-	},
-	"Bm": {
-		"standard": [
-			"B4",
-			"D5",
-			"Gb5"
-		],
-		"spread": [
-			"B3",
-			"D5",
-			"Gb6"
-		],
-		"rich": [
-			"B2",
-			"B3",
-			"D5",
-			"Gb5",
-			"B5"
-		]
-	}
-}
